Versión 3.2 -creada FloatTest.java -creado ttBinary() en FloatTest.java -añadida documentación a Binary.java  -suma1()  -sign()  -flip()  -negative() -añadido getValue() a Float.java -modificado hashCode() para Float.java -modificado equals() para Float.java -modificado ttBinary() en Float.java -modificado ttFloat() en Int.java -creado DecimalToBinary() en Num.java -añadida documentación a Num.java  -DecimalToBinary()  -reverse()  -PosIntToBinary()  -NegIntToBinary()

+++ src/main/java/cl/uchile/dcc/scrabble/gui/tarea1/Int.java
@@ -37,7 +37,7 @@ public class Int extends Num implements TypesInterface{
     /* Chequear si funciona*/
     @Override
     public Float ttFloat() {
-        return new Float(this.value);
+        return new Float((float) this.value);
     }
 
     @Override
